Module parse::ast

Types and traits for working with (abstract) syntax trees.

This includes the enum AstNode and the struct Ast, as well as traits such as Walker for walking a tree.

Ast, AstNode, and (Abstract) Syntax Trees

Abstract syntax trees and concrete syntax trees are represented through Ast and AstNode. AstNode represents a node in a syntax tree, either a leaf containing a terminal, or a branch containing an Ast. An instance of Ast represents a branch node in a syntax tree, with a label and one or more child AstNode’s.

In an Ast<N, T>, the generic types N and T recursively determine the type of labels on branch nodes and the type of leaf nodes, respectively, on the tree as a whole; the label of an Ast<N, T> is of type N, and the children are of type AstNode<N, T>, where the leafs of an AstNode<N, T> are of type T and the branches are of type Ast<N, T>

Walking an Ast

This module also contains facilities for recursively walking an Ast. These consist of a single general pattern with three variations differing based on ownership: borrowing, mutable borrowing, and consuming. For each degree of ownership, there are two items:

A walker trait, containing three methods: on_branch, to be called when entering a branch node, on_leaf, to be called on leaf nodes, and exit, to be called when exiting a branch node.

And a walking iterator struct that adapts a Walker and implements Iterator by recursively walking a tree and returning the results of calling the appropriate methods on the Walker.

The different items for each access type/role combo are illustrated on the following table.

Trait	Iter Adapter

Borrowing:	Walker	WalkerIter

Mut Borrowing:	WalkerMut	WalkerIterMut

Consuming:	Consumer	WalkerIntoIter

For the walker trait, the methods return options; if the returned variant is None, the iterator adapter will skip it and move on to the next node.1

The on_branch method is also used as a guard by returning a tuple where the second item is a bool2. If the bool is false, the iterator adapter does not descend for that branch, and skips to the next node under the current parent branch3.

Differences

There are minor differences between the three variations, having to do with the different constraints (or lack thereof) on the differeing degrees of ownership. Namely:

For WalkerMut, the mutable walker trait, all three methods accept an additional parent parameter, which is a reference to the label of the parent node. This helps alleviate the need for the WalkerMut to alias references to the label’s to keep track of state.

For Consumer<A, T>, the consuming walker trait, the method on_branch takes ownership of the branch node and returns (Option<T>, Option<A>) instead of (Option<T>, bool). If the Option<A> is None, that node and its descendents are skipped, just like returning false for the other traits. If the Option<A> is [Some(ast)], then the iterator adapter (WalkerIntoIter) recursively descends along that tree.

For Walker, the borrowing walking trait, the method exit takes a reference to the branch node as a whole, while the other two traits only act on the label of the branch being exited.

Note that the “next node” is the one it would have gone to on a subsequent call to next. In other words, if on_branch returns None on a node n, the iterator adapter will descend and move on to the first child node of n, rather than going to the next sibling node. For skipping a branch node and all its descendants, see the next paragraph. ↩

With a slight exception for Consumer. See the documentation there and in the next section for more details. ↩

Note that the boolean value only controls whether or not the iterator descends for that branch node; if the Option result is Some, the iterator still returns that result, and simply moves on to the next sibling node on the next call to Iterator::next. ↩

Structs

Ast

This is the trait for branch nodes in a (abstract) syntax tree.

DisplayTree

A wrapper for an Ast that implements Display by recursively printing the labels of the branch node and the leaf nodes of its tree

IntoIter

An Iterator over the labels and leaf nodes of an Ast.

Iter

An Iterator by reference over the nodes of an Ast.

IterMut

An Iterator by mutable reference over the labels and leaf nodes of an Ast.

WalkerIntoIter

A struct for walking an Ast while taking ownership of it.

WalkerIter

A struct for walking an Ast by reference.

WalkerIterMut

A struct for walking an Ast by mutable reference.

Enums

AstElem

A component of an Ast: either a label or a leaf.

AstNode

A node in an (abstract) syntax tree.

MutNode

A component of an Ast: either a label or a leaf.

MutRefNode

A mutably borrowing adapter for AstNode.

RefNode

A borrowing adapter for AstNode.

Traits

Consumer

A trait for owning walkers.

Walker

A trait for borrowing walkers.

WalkerMut

A trait for mutably borrowing walkers.

Functions

reverse

Recursively reverses the order of all the child nodes of an Ast and of all sub-trees.