Struct parse::ast::Ast

pub struct Ast<N, T> { /* fields omitted */ }

This is the trait for branch nodes in a (abstract) syntax tree.

The generic types N and T determine the nature of the full tree rooted at any given instance of Ast<N, T>: N determines the labels used for branch nodes (i.e. the non-terminal symbols,) typically an enum of some kind, and T determines the contents of the leaf nodes, typically, but not necessarily, Token<'_, K> for some choice of terminal discriminant K.

Label and Children

Generally speaking, there are two components to an Ast<N, T>: The label, which is an instance of N and describes what kind of branch node self is, and the children, which is a vec of AstNode<N, T> representing the child nodes to self.

Implementations

impl<N, T> Ast<N, T>

pub fn label(&self) -> &N

Returns a reference to the label of self.

pub fn label_mut(&mut self) -> &mut N

Returns a mutable reference to the label of self.

pub fn into_label(self) -> N

Consumes self and returns the label.

pub fn children(&self) -> &[AstNode<N, T>]

Returns a reference to the children of self.

pub fn children_mut(&mut self) -> &mut Vec<AstNode<N, T>>

Returns a mutable referenct to the children of `self.

pub fn into_children(self) -> Vec<AstNode<N, T>>

Consumes self and returns the children.

pub fn components(&self) -> (&N, &[AstNode<N, T>])

Returns a reference to the label and children of self.

pub fn components_mut(&mut self) -> (&mut N, &mut Vec<AstNode<N, T>>)

Returns a mutable reference to the label and children of self.

pub fn into_components(self) -> (N, Vec<AstNode<N, T>>)

Consumes self the label and children.

pub fn walk<'a, W, R>(&'a self, walker: W) -> WalkerIter<'a, N, T, R, W> where
W: Walker<'a, N, T, R>,

Takes a Walker and returns a WalkerIter for walking self.

pub fn walk_default<'a, W, R>(&'a self) -> WalkerIter<'a, N, T, R, W> where
W: Walker<'a, N, T, R> + Default,

Like walk but uses the default value of W.

pub fn walk_mut<'a, W, R>(
&'a mut self,
walker: W
) -> WalkerIterMut<'a, N, T, R, W> where
W: WalkerMut<'a, N, T, R>,

Takes a WalkerMut and returns a WalkerIterMut for mutably walking self.

pub fn walk_mut_default<'a, W, R>(&'a mut self) -> WalkerIterMut<'a, N, T, R, W> where
W: WalkerMut<'a, N, T, R> + Default,

Like walk_mut but uses the default value of W.

pub fn consume<C, R>(self, consumer: C) -> WalkerIntoIter<N, T, R, C> where
C: Consumer<N, T, R>,

Takes a Consumer and returns a WalkerIntoIter for walking self while taking ownership of the nodes.

pub fn consume_default<C, R>(self) -> WalkerIntoIter<N, T, R, C> where
C: Consumer<N, T, R> + Default,

Like consume but uses the default value of W.

pub fn new(label: N, children: Vec<AstNode<N, T>>) -> Self

Creates a new instance of Self with the given label and children

pub fn push(&mut self, node: AstNode<N, T>)

Pushes a new node onto the end of the children of self.

pub fn push_branch(&mut self, branch: Self)

Pushes a new branch node onto the end of the children of self.

Equivalent to self.push(AstNode::Branch(branch).

pub fn push_leaf(&mut self, leaf: T)

Pushes a new leaf node onto the end of the children of self.

Equivalent to self.push(AstNode::Leaf(leaf).

pub fn display(&self) -> DisplayTree<'_, N, T> where
T: Display,
N: Display,

Wraps self in a DisplayTree for printing.

pub fn iter(&self) -> Iter<'_, N, T>

Returns an iterator over references to the nodes of self.

pub fn iter_mut(&mut self) -> IterMut<'_, N, T>

Returns an iterator over mutable references to the labels and leafs of self.

pub fn into_iter(self) -> IntoIter<N, T>

Returns an iterator over the labels and leafs of self.