Trait parse::ast::Walker

pub trait Walker<'a, B, L, T> {
    fn on_branch(&mut self, tree: &'a Ast<B, L>) -> (Option<T>, bool);

    fn on_leaf(&mut self, leaf: &'a L) -> Option<T>;

    fn exit(&mut self, tree: &'a Ast<B, L>) -> Option<T>;
}

A trait for borrowing walkers.

Used in conjunction with WalkerIter, which takes a Walker and an Ast and implements Iterator by recursively walking the Ast and (lazily) calling the associated methods.

If a value of None is returned, the WalkerIter skips that node and moves on to the next one.

if the second value returned by on_branch is false, the WalkerIter does not descend to the child nodes of that branch (it will still return the value for the current node if the Option was [Some(t)].)

The three methods of this trait are called by the WalkerIter at different points when walking the tree:

on_branch is called on a branch (non-leaf) node before descending to its child nodes.

on_leaf is called on a leaf node.

exit is called upon exiting a branch node, after having covered all of its children. It will not be called if on_branch returned false for that branch.

See the module documentation for more.

Required methods

fn on_branch(&mut self, tree: &'a Ast<B, L>) -> (Option<T>, bool)

Called when entering a branch node.

If the boolean value is false, the WalkerIter does not descend for that branch node, and skips all of its child nodes.

If the Option value is None, that node is skipped. Otherwise the WalkerIter returns the value contained in the Some variant.

fn on_leaf(&mut self, leaf: &'a L) -> Option<T>

Called on leaf nodes.

If the return value is None, that node is skipped. Otherwise the WalkerIter returns the value contained in the Some variant.

fn exit(&mut self, tree: &'a Ast<B, L>) -> Option<T>

Called on branch nodes after having covered all of their child nodes.

If on_branch returned false for that branch node, this method is not called.

If the return value is None, that node is skipped. Otherwise the WalkerIter returns the value contained in the Some variant.