Macro parse::build_parser

macro_rules! build_parser {
    ($(name = $name : expr $(, location = $location : expr) ? $(,) ?) ?) => { ... };
    (location = $location : expr $(,) ?) => { ... };
    ($(name = $name : expr,) ? $(location = $location : expr,) ? discriminant =
 $discriminant : ty $(,) ?) => { ... };
    ($(name = $name : expr,) ? $(location = $location : expr,) ? N = $N : ty $(,)
 ?) => { ... };
    ($(name = $name : expr,) ? $(location = $location : expr,) ? T = $T : ty $(,)
 ?) => { ... };
    ($(name = $name : expr,) ? $(location = $location : expr,) ? N = $N : ty, T =
 $T : ty $(,) ?) => { ... };
}

A convenience macro for building parsers in a build script.

This macro can optionally take, in order, a name, location, non-terminal discriminant type N, and terminal discriminant type T1. If N or T aren’t specified, they default to u32. name and location have the same defaults as described in ParserBuilder.

name should be of a type that implements Into<OsString>, and location should be of a type that implements Into<PathBuf>, as described at ParserBuilder::with_name and ParserBuilder::at_location, respectively.

It may also take a discriminant type in place of both N and T, which, if specified, acts as the discriminant for both non-terminals and terminals.

Arguments are given in the form var = $var, seperated by commas. For example:

// in build.rs

parse::build_parser!(name = "my_parser", T = u8);

This macro returns a Result<ParserBuilder, ParserError> containing either the ParserBuilder used in constructing the parser, or any ParserError’s incurred.

// in build.rs

use std::fs;

fn main() {
	let source = fs::read_to_string("my_parser.y")
		.expect("error reading 'my_parser.y'");

	let builder = parse::build_parser!(location = "my_parser.y")
		.expect("error processing parser file 'my_parser.y'");

	assert_eq!(source, builder.src().unwrap());
}

Panic

Since this macro is essentially a wrapper for ParserBuilder, it panics when not being run in a build script. See the documentation on ParserBuilder::build_parser or here for more details.

// in src/main.rs

parse::build_parser!();

Reruns

This macro automatically runs ParserBuilder::set_rerun to set cargo to watch the location for changes and rerun the build script if it notices any.

Note that this also means that the build script will not be rerun unless cargo detects any changes in a file it has been specifically told to watch. Keep this in mind if you have other components to the build script than building a parser. See the documentation for more.

Examples

Basic usage and default arguments.

// builds a parser with name "my_parser", location "parser.y", non-terminal
// discriminant u8, and terminal discriminant u16
parse::build_parser!(name = "my_parser", location = "parser.y", N = u8, T = u16)?;

// builds a parser with name "alt_parser", default location (in this instance,
// "alt_parser.y",) default non-terminal discriminant of u32, and terminal
// discriminant u8
parse::build_parser!(name = "alt_parser", T = u8)?;
Ok(())

Using discriminant.

// builds a parser with default name and location, and with u16 for both discriminant
// types
// parse::build_parser!(N = u16, T = u16)?;

// same thing as above, but more concise
parse::build_parser!(discriminant = u16)?;
Ok(())

It is an error to give the arguments out of order.

parse::build_parser!(location = "parser.y", name = "my_parser")?;
Ok(())

It is also an error to use discriminant with N or T.

parser::build_parser!(discriminant = u8, N = u8)?;
Ok(())

See the documentation on ParserBuilder for more information on what name and location do, and what their requirements are. ↩