Struct parse::ParserBuilder

pub struct ParserBuilder { /* fields omitted */ }

A helper struct for creating parsers using build scripts

Instances of this stuct can be used in build scripts for creating parsers to be used in a package. It has two hidden paramaters that affect its behaviour: the name used by the builder when writing the finished parser, and the location of the source file. These are private and can only be accessed or modified through the provided methods which automatically run correctness checks on the provided values.

Note that ParserBuilder’s are only intended to be run by Cargo as a part of a build script, and ParserBuilder::new will not return a builder unless the environment resembles that of a build script. See the official documentation for more details on build scripts.

Name and Location

A ParserBuilder will write the parser in the current builds OUT_DIR, with the name [name].rs. name must be a valid filename, with no seperators1 and no extension. If name is not set, but location is, name defaults to the filename for location minus the “.y” extension. Otherwise it defaults to the name of the package with “_parse” added to the end.

The value for location must be an existing file ending with “.y”. If location is not set then it defaults to name with “.y” added to the end (if name and location are not set, name default to the package name plus “_parse”, so location would default to "[package name]_parse.y`“.)

Environment

ParserBuilder’s rely on a number of environment variables to be set in order to understand the package being built and where to write files. These should be automatically set when the code is being run by Cargo as a part of a build script. For safety, ParserBuilder::new() will check the environment and will not return a builder unless it is properly configured.

///outside of build.rs

use parse::ParserBuilder;

assert!(ParserBuilder::new().is_none());

Reruns

Once the desired location for the source file of a parser is set it is recommended to run set_rerun to set Cargo to monitor the source file for changes and rerun the build script when it detects one since the last build.

“/” on Unix or Redox, and “\” on Windows. Note that, since this is running as part of a build script, it is the host system that matters, not the compilation target. This also means that you should be weary about using either / or \, regardless of the platform: something that works fine on your system may fail to compile when being built on a different platform! ↩

Implementations

impl ParserBuilder

pub fn new() -> Option<Self>

Creates a new instance of a ParserBuilder. Returns None if the environement is not configured properly. See here for more details.

Example

let mut builder = ParserBuilder::new().unwrap();
builder.set_default_values();

pub fn name(&self) -> Option<&OsStr>

Returns the name of self if set. Returns None otherwise.

pub fn location(&self) -> Option<&Path>

Returns the location of self if set. Returns None otherwise.

pub fn src(&self) -> Option<&str>

Returns the source string used by the builder to generate a parser, which is read from self.location().

Returns Some(src) if a source string has been loaded, and None otherwise.

Note that, even if the location for self is set, this method may still return None if the location has not been loaded yet. For a version of this method that automatically loads the source, see load_file.

Example

use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();

builder
	.at_location("my_parser.y")?
	.load_file()?;

let source = builder.src().unwrap();

pub fn with_name<S: Into<OsString>>(
&mut self,
name: S
) -> Result<&mut Self, BuilderError>

Attempts to set the name for self to name.

If name is a valid filename (see here,) then this method returns a mutable reference to self, otherwise it returns an BuilderError.

Example

use std::ffi::OsStr;
use parse::ParserBuilder;
let mut builder = ParserBuilder::new().unwrap();
let builder_name = builder
	.with_name("my_parser")?
	.name();
assert_eq!(builder_name, Some(OsStr::new("my_parser")));

pub fn at_location<S: Into<PathBuf>>(
&mut self,
location: S
) -> Result<&mut Self, BuilderError>

Attempts to set the location for self to location.

If location is a valid path to a source file (see here,) then this method returns a mutable reference to self, otherwise it returns an BuilderError.

Example

use std::path::Path;
use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();
let builder_location = builder
	.at_location("lex-yacc/my_parser.y")?
	.location();

assert_eq!(builder_location, Some(Path::new("lex-yacc/my_parser.y")));

pub fn set_default_values(&mut self)

Sets the name and location to their default values, as described here. Calling this method directly is usually unecessary, as the fields will automatically be set to their default values as necessary when building a parser. It can however be usefull for inspecting what the values would be.

Example

use std::ffi::OsStr;
use std::path::Path;
use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();
builder.with_name("my_parser").unwrap();

assert_eq!(builder.location(), None);

builder.set_default_values();

assert_eq!(builder.name(), Some(OsStr::new("my_parser")));
assert_eq!(builder.location(), Some(Path::new("my_parser.parse")));

pub fn set_rerun(&self) -> Option<()>

Tells Cargo to watch the source file for changes and rerun the build script if necessary. If no location is set, it returns None

Example

// in build.rs

use std::error::Error;
use parse::ParserBuilder;

fn main() -> Result<(), Box<dyn Error>> {
	let mut builder = ParserBuilder::new().unwrap();
	builder
		.with_name("my_parser")
		.unwrap()
		.build::<u32, u32>()?
		.set_rerun().unwrap();
	Ok(())
}

pub fn force_load_file(&mut self) -> Result<&str, IOError>

Like load_file, except it always reads the location, even if a source text has already been loaded.

This may be usefull if, for instance, the location for self has been changed since loading a source file, and you want to make sure that the new location is loaded.

Example

use std::fs;
use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();

let old_source = builder
	.at_location("my_parser.y")?
	.load_file()?
	.to_string();

let new_source = fs::read_to_string("new_parser.y")?;

builder.at_location("new_parser.y")?;

assert_eq!(old_source, builder.load_file().unwrap());	// load_file returns the cached
														// content of "my_parser.y",
														// instead of "new_parser.y"

assert_eq!(new_source, builder.force_load_file()?);		// loads and returns the content
														// of "new_parser.y"

assert_eq!(new_source, builder.load_file().unwrap());	// load_file now returns the
														// cached content of
														// "new_parser.y"

pub fn load_file(&mut self) -> Result<&str, IOError>

Loads and returns the text of the source file at the location for self.

If the source file has already been loaded, this method returns the cached text of that file, rather than loading it again. For the version that reloads the source file, see force_load_file.

Example

use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();

let location_content = builder
	.at_location("my_parser.y")?
	.load_file()?
	.to_string();

let source = builder.src().unwrap();

assert_eq!(location_content, source);

pub fn scan_src<N: Prim + Discriminant, T: Prim + Discriminant>(
&self
) -> Option<Result<Rules<'_, N, T>, ParserError>>

Scans self.src() and attempts to parse it into a Rules.

If self.src() is None, then this function returns None, otherwise it returns the Result of parsing self.src().

Example

use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();

builder.load_file()?;

let rules = builder.scan_src::<u32, u32>().unwrap()?;

pub fn get_rules<N: Prim + Discriminant, T: Prim + Discriminant>(
&mut self
) -> Result<Rules<'_, N, T>, ParserError>

Like scan_src except it loads a source from the location if self.src() is None.

Example

use parse::ParserBuilder;

let mut builder = ParserBuilder::new().unwrap();

let rules = builder.get_rules::<u32, u32>()?;

pub fn build<N, T>(&mut self) -> Result<&mut Self, ParserError> where
N: Prim + Discriminant,
T: Prim + Discriminant,

Builds the parser, reading from the file at location and writing a file with the appropriate name to the out dir set by Cargo.

Example

///in build.rs

use std::error::Error;
use parse::ParserBuilder;

fn main() -> Result<(), Box<dyn Error>> {
	let mut builder = ParserBuilder::new().unwrap();
	builder
		.build::<u32, u32>()?
		.set_rerun().unwrap();
	Ok(())
}

pub fn build_parser<N, T>() -> Result<Self, ParserError> where
N: Prim + Discriminant,
T: Prim + Discriminant,

Builds a parser using the default name and location.

Equivalent to calling ParserBuilder::new() and then calling build on the result.

Panic

This function panics if not being run by cargo in a build script.

use parse::ParserBuilder;

ParserBuilder::build_parser::<u32, u32>();