Struct parse::Rules

pub struct Rules<'a, N: Prim, T: Prim> { /* fields omitted */ }

A struct representing a parsed collection of grammar or production rules.

'a is the lifetime of the source text from which the rules were parsed, and N and T are the discriminant types for non-terminal and terminal symbols respectively.

Creating Rules

Instances of Rules can be created with Rules::new, which parses a source str and returns its Rules.

use parse::Rules;

let src = r"
	Start -> Expr;
	Expr -> Fact;
	Fact -> lparen Expr rparen;
	Fact -> n;
	Fact -> plus Fact;
	Fact -> Fact plus n;
";

let rules = Rules::<u32, u32>::new(src).unwrap();

Alternatively, one can use Rules::from_rule_map to create a Rules from a mapping of strings for non-terminals to the bodies of their rules. For example:

use std::collections::HashMap;
use parse::Rules;

let mut map: HashMap<&str, Vec<Box<[&str]>>> = HashMap::new();

map.insert("Start", vec![["Expr"].into()]);
map.insert("Expr", vec![["Fact"].into()]);
map.insert("Fact", vec![
	["lparen", "Expr", "rparen"].into(),
	["n"].into(),
	["plus", "Fact"].into(),
	["Fact", "plus", "n"].into()
]);

let rules = Rules::<u32, u32>::from_rule_map(&map).unwrap();

Using Rules

Rules contains all the information necessary for generating a parser for the language defined by the given production rules. The process of generating and writing a parser module to a writer can be streamlined with Rules::write. However, the process of generating a parser is computationally expensive. So if you intend to use the generated parser tables multiple times, it is recomended to use a ParserSpec, which stores the result of generating a parser.

use parse::Rules;

let src = r"
	Start -> Expr;
	Expr -> Fact;
	Fact -> lparen Expr rparen;
	Fact -> n;
	Fact -> plus Fact;
	Fact -> Fact plus n;
";

let rules = Rules::<u32, u32>::new(src).unwrap();

let mut fout_1: Vec<u8> = Vec::new();
let mut fout_2: Vec<u8> = Vec::new();
let mut fout_3: Vec<u8> = Vec::new();

rules.write(&mut fout_1).unwrap();	// Generates, then consumes a ParserSpec

let spec = rules.gen_spec();

spec.write(&mut fout_2).unwrap();
spec.write(&mut fout_3).unwrap();	// Avoids re-generating a ParserSpec

Implementations

impl<'a, N, T> Rules<'a, N, T> where
N: Prim + Discriminant,
T: Prim + Discriminant,

pub fn from_rule_map(
map: &HashMap<&'a str, Vec<Box<[&'a str]>>>
) -> Option<Result<Self, InvalidDiscriminant>>

Creates a Rules from a representation of a set of grammar rules in the form of a mapping from the non-terminal symbols (as str’s) to Vec’s of their right hand sides (as boxed slices of str’s.)

Returns Some(result) if map is a valid grammar specification, otherwise it returns None.

Example

use std::collections::HashMap;
use parse::Rules;

let mut map: HashMap<&str, Vec<Box<[&str]>>> = HashMap::new();

map.insert("Start", vec![["Expr"].into()]);
map.insert("Expr", vec![["Fact"].into()]);
map.insert("Fact", vec![
	["lparen", "Expr", "rparen"].into(),
	["n"].into(),
	["plus", "Fact"].into(),
	["Fact", "plus", "n"].into()
]);

let rules = Rules::<u32, u32>::from_rule_map(&map).unwrap();

pub fn new<S: AsRef<str> + ?Sized>(src: &'a S) -> Result<Self, ParserError>

Parses a grammar specification src into a Rules

Example

use parse::Rules;

let src = r"
	Start -> Expr;
	Expr -> Fact;
	Fact -> lparen Expr rparen;
	Fact -> n;
	Fact -> plus Fact;
	Fact -> Fact plus n;
";

let rules = Rules::<u32, u32>::new(src).unwrap();

impl<'a, N, T> Rules<'a, N, T> where
N: Prim + Discriminant,
T: Prim + Discriminant,

pub fn write<W: Write>(&self, fout: W) -> Result<()>

Generates a ParserSpec and uses it to write a parser module to fout.

Example

use std::fs::File;
use parse::Rules;

let src = r"
	Start -> Expr;
	Expr -> Fact;
	Fact -> lparen Expr rparen;
	Fact -> n;
	Fact -> plus Fact;
	Fact -> Fact plus n;
";

let rules = Rules::<u32, u32>::new(src).unwrap();

let fout = File::create("parser.rs")
	.expect("unable to create file 'parser.rs'");

rules.write(fout)
	.expect("unable to write to 'parser.rs'");

impl<'a, N: Prim, T: Prim> Rules<'a, N, T>

pub fn gen_spec(&self) -> ParserSpec<'a, '_, N, T>

Generates and returns a ParserSpec based on the rules provided by self.

Example

use std::fs::File;
use parse::{Rules, ParserSpec};

let src = r"
	Start -> Expr;
	Expr -> Fact;
	Fact -> lparen Expr rparen;
	Fact -> n;
	Fact -> plus Fact;
	Fact -> Fact plus n;
";

let rules = Rules::new(src).unwrap();

let spec: ParserSpec<'_, '_, u32, u32> = rules.gen_spec();