Trait parse::FromUsize

pub trait FromUsize {
    const MAX: usize;

    fn try_from(n: usize) -> Option<Self>
    where
        Self: Sized;

    fn from(n: usize) -> Self
    where
        Self: Sized,
    { ... }
}

A trait for converting from usize to a discriminant. This is required for generating a Rules or ParserSpec from a source text.

A type that implements both FromUsize and Eq must ensure that the implementation of from must preserve and reflect equality, e.g.:

use parse::FromUsize;

fn test_eq<T: Eq + FromUsize>(l: usize, r: usize) {
	let is_eq_int = l == r;
	let is_eq = T::from(l) == T::from(r);
	assert_eq!(is_eq_int, is_eq);	// l == r if and only if T::from(l) == T::from(r)
}

Implementing both FromUsize and IntoUsize

If a type implements both FromUsize and IntoUsize, it should ensure that from and into are isomorphisms. That is, calling from and then into should return the same number as the one you started with (or panic,) and calling into and then from should return an equivalent object to the original (or panic.) For example:

use parse::{FromUsize, IntoUsize};

fn endo_t<T: FromUsize + IntoUsize>(t: T) -> T {
	T::from(t.into())
}

fn endo_usize<T: FromUsize + IntoUsize>(n: usize) -> usize {
	T::from(n).into()
}

assert_eq!(128u8, endo_t(128u8));
assert_eq!(128, endo_usize::<u8>(128));

Associated Constants

const MAX: usize

The maximum size that Self can be converted from.

Required methods

fn try_from(n: usize) -> Option<Self> where
Self: Sized,

Converts a usize to a Self. Returns None if n is greater than Self::MAX, otherwise returns Some. For exmple:

use parse::FromUsize;

assert_eq!(<u8 as FromUsize>::MAX, 255);

let none = <u8 as FromUsize>::try_from(256);
let some = <u8 as FromUsize>::try_from(255);

assert!(matches!(none, None));
assert!(matches!(some, Some(255)));

Provided methods

fn from(n: usize) -> Self where
Self: Sized,

Equivalent to Self::try_from(n).unwrap().

Panic

Implementations should panic if n is greater than Self::MAX. Otherwise, the result should be the same as calling unwrap on the result of Self::try_from.