Trait parse::IntoUsize

pub trait IntoUsize {
    fn try_into(self) -> Option<usize>
    where
        Self: Sized;

    fn into(self) -> usize
    where
        Self: Sized,
    { ... }
}

A trait for converting from a discriminant to usize. This is required for writing a parser using a ParserSpec.

A type that implements both IntoUsize and Eq must ensure that the implementation of into must preserve and reflect equality, e.g.:

use parse::IntoUsize;

fn test_eq<T: Eq + IntoUsize>(l: T, r: T) {
	let is_eq = &l == &r;
	let is_eq_int = l.into() == r.into();
	assert_eq!(is_eq, is_eq_int);	// l == r if and only if l.into() == r.into()
}

Implementing both FromUsize and IntoUsize

If a type implements both FromUsize and IntoUsize, it should ensure that from and into are isomorphisms. That is, calling from and then into should return the same number as the one you started with (or panic,) and calling into and then from should return an equivalent object to the original (or panic.) For example:

use parse::{FromUsize, IntoUsize};

fn endo_t<T: FromUsize + IntoUsize>(t: T) -> T {
	T::from(t.into())
}

fn endo_usize<T: FromUsize + IntoUsize>(n: usize) -> usize {
	T::from(n).into()
}

assert_eq!(128u8, endo_t(128u8));
assert_eq!(128, endo_usize::<u8>(128));

Required methods

fn try_into(self) -> Option<usize> where
Self: Sized,

Converts self into a usize.

If Self implements both FromUsize and IntoUsize, then this function should always return Some on a value that can be generated by FromUsize.

use parse::{FromUsize, IntoUsize};

let n = <u8 as FromUsize>::from(128);

let result = <u8 as IntoUsize>::try_into(n);

assert!(matches!(result, Some(_)));

See also the requirements here for a type that implements both FromUsize and IntoUsize.

Provided methods

fn into(self) -> usize where
Self: Sized,

Equivalent to calling self.try_into().unwrap().

Panic

Implementations should panic when try_into would return None. Otherwise, they should return the equivalent of calling unwrap on [self.try_into].