Trait parse::Parser

pub trait Parser<'a, L, T, B, U>: Sized where
    L: Tokens<'a, T>, {
    fn new<I>(inp: I) -> Result<Self, Error>
    where
        I: IntoIterator<IntoIter = L>;

    fn parse(&mut self) -> Result<Ast<B, U>, Error>;
}

A trait providing a generic interface for Parsers.

new creates a new Parser from a token stream, and parse processes the token stream into an Ast.

These methods can also be found in the public interfaces of the parser types generated by this crate. This means that importing this trait is often not necessary when using Parsers.

Required methods

fn new<I>(inp: I) -> Result<Self, Error> where
I: IntoIterator<IntoIter = L>,

fn parse(&mut self) -> Result<Ast<B, U>, Error>